# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Statebound HTTP API
  version: "0.8.0"
  summary: Read-only governance API for Statebound core (Phase 8 wave B).
  description: |
    The Statebound HTTP API exposes the core's read surface so the optional
    `statebound-reason` add-on (and other registered ServiceAccount clients)
    can consume it.

    ## Boundaries

    - **Read-only.** Write operations (approve, plan, apply) remain CLI-only
      in v0.8. Agents never mutate approved state.
    - **Authenticated.** Every `/v1/...` route requires a valid bearer token.
      Tokens may be OIDC-issued JWTs (when configured with
      `--oidc-issuer`/`--oidc-audience`) or a static dev token (when
      configured with `--dev-token`).
    - **OPA-gated.** Authorization is evaluated through the same RBAC layer
      used by the CLI; the `product:read` capability gates every read route.
    - **No private material.** Signing keys expose only the public half plus
      fingerprint. `private_key_ref` is never returned.

  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development.
security:
  - bearerAuth: []
...
paths:
...
  /v1/products:
    get:
      tags: [products]
      summary: List products.
      operationId: listProducts
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedProducts" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/products/{id}:
    get:
      tags: [products]
      summary: Get a product by id.
      operationId: getProduct
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/plans:
    get:
      tags: [plans]
      summary: List plans.
      operationId: listPlans
      parameters:
        - in: query
          name: product_id
          schema: { type: string }
        - in: query
          name: approved_version_id
          schema: { type: string }
        - in: query
          name: connector_name
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedPlans" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/plans/{id}:
    get:
      tags: [plans]
      summary: Get a plan with items.
      operationId: getPlan
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlanDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/policy-decisions:
    get:
      tags: [policy]
      summary: List policy decisions.
      operationId: listPolicyDecisions
      parameters:
        - in: query
          name: change_set_id
          schema: { type: string }
        - in: query
          name: outcome
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedPolicyDecisions" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/policy-decisions/{id}:
    get:
      tags: [policy]
      summary: Get a policy decision.
      operationId: getPolicyDecision
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PolicyDecision" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/signing-keys:
    get:
      tags: [signing]
      summary: List signing keys (public material only).
      operationId: listSigningKeys
      parameters:
        - in: query
          name: include_disabled
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedSigningKeys" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/signing-keys/{key_id}:
    get:
      tags: [signing]
      summary: Get a signing key (public material only).
      operationId: getSigningKey
      parameters:
        - in: path
          name: key_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SigningKey" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC-issued JWT (production) or a static dev token (local dev). The
        token's `sub` claim (or the dev-actor mapping) is resolved against
        the actor_role_bindings table to determine RBAC capabilities.
  parameters:
    PathID:
      in: path
      name: id
      required: true
      schema: { type: string }
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 0, maximum: 1000, default: 100 }
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but lacks the required capability.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "missing_token"
        message:
          type: string
          example: "Authorization header is required"
        request_id:
          type: string
    Actor:
      type: object
      required: [kind, subject]
      properties:
        kind:
          type: string
          enum: [human, service_account, system]
        subject:
          type: string
...
    Product:
      type: object
      required: [id, name, owner, created_at, updated_at]
      properties:
        id: { type: string }
        name: { type: string }
        owner: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PaginatedProducts:
      type: object
      required: [items, count]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        count: { type: integer }
        next_cursor:
          type: string
          nullable: true
...
    Plan:
      type: object
      required:
        - id
        - product_id
        - approved_version_id
        - sequence
        - connector_name
        - connector_version
        - state
        - content_hash
        - generated_at
        - generated_by
      properties:
        id: { type: string }
        product_id: { type: string }
        approved_version_id: { type: string }
        sequence: { type: integer }
        connector_name: { type: string }
        connector_version: { type: string }
        state:
          type: string
          enum: [draft, ready, refused, applied, failed]
        summary: { type: string }
        content_hash: { type: string }
        generated_at: { type: string, format: date-time }
        generated_by: { $ref: "#/components/schemas/Actor" }
        refused_reason: { type: string }
    PlanItem:
      type: object
      required: [id, plan_id, sequence, action, resource_kind, resource_ref]
      properties:
        id: { type: string }
        plan_id: { type: string }
        sequence: { type: integer }
        action: { type: string }
        resource_kind: { type: string }
        resource_ref: { type: string }
        body:
          type: object
          additionalProperties: true
        risk: { type: string }
        note: { type: string }
    PlanDetail:
      allOf:
        - $ref: "#/components/schemas/Plan"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/PlanItem" }
    PaginatedPlans:
      type: object
      required: [items, count]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Plan" }
        count: { type: integer }
        next_cursor:
          type: string
          nullable: true
...
    PolicyDecision:
      type: object
      required:
        - id
        - change_set_id
        - phase
        - outcome
        - bundle_hash
        - evaluated_at
      properties:
        id: { type: string }
        change_set_id: { type: string }
        phase: { type: string }
        outcome: { type: string }
        rules:
          type: array
          items:
            type: object
            additionalProperties: true
        input:
          type: object
          additionalProperties: true
        bundle_hash: { type: string }
        evaluated_at: { type: string, format: date-time }
    PaginatedPolicyDecisions:
      type: object
      required: [items, count]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/PolicyDecision" }
        count: { type: integer }
        next_cursor:
          type: string
          nullable: true
    SigningKey:
      description: |
        A signing key. Only the public material (public_key bytes,
        fingerprint, lifecycle metadata) is exposed. The `private_key_ref`
        and any private bytes are NEVER included in API responses.
      type: object
      required:
        - key_id
        - algorithm
        - public_key
        - fingerprint
        - created_at
        - created_by
        - disabled
      properties:
        key_id: { type: string }
        algorithm: { type: string, enum: [ed25519] }
        public_key:
          description: Base64-encoded public-key bytes.
          type: string
        fingerprint: { type: string }
        created_at: { type: string, format: date-time }
        created_by: { $ref: "#/components/schemas/Actor" }
        expires_at:
          type: string
          format: date-time
          nullable: true
        disabled: { type: boolean }
        note: { type: string }
        last_used_at:
          type: string
          format: date-time
          nullable: true
    PaginatedSigningKeys:
      type: object
      required: [items, count]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SigningKey" }
        count: { type: integer }
        next_cursor:
          type: string
          nullable: true
...